Select between install everything for autonomous monitoring, or just a user computer

+++ config/install.sh
@@ -1,26 +1,39 @@
-#!/bin/bash
+#!/usr/bin/env bash
+
+if [ "$#" -ne 1 ]; then
+    echo "Usage: $./install.sh [install_monitoring_system=0,1]"
+    exit 1
+fi
+
+if [[ $1 -eq 1 ]] ; then
+    echo "Starting install install script for monitoring systems!"
+fi
+
 set -ex
 working_dir=`pwd`
-
 # Create dependencies directory
 mkdir -p ~/dependencies
 mkdir -p ~/code
-[ -f ~/dependencies/pats_ssh_files.tar.xz ] || {
-	mv pats_ssh_files.tar.xz ~/dependencies
-}
-pushd ~/dependencies
 
-#Change hostname: sudo hostname pats-proto1 
-#Update: sudo nano /etc/hosts 
-#Change 127.0.1.1 pats-proto1 
-#Make it persistant, update: sudo nano /etc/hostname
-if [[ $HOSTNAME == pats-proto* ]];
-then
-  echo "Hostname changed detected, good."
-else
-  echo "Change hostname before running this script!"
-  exit 1
+if [[ $1 -eq 1 ]] ; then
+    [ -f ~/dependencies/pats_ssh_files.tar.xz ] || {
+        mv pats_ssh_files.tar.xz ~/dependencies
+    }
+
+    #Change hostname: sudo hostname pats-proto1 
+    #Update: sudo nano /etc/hosts 
+    #Change 127.0.1.1 pats-proto1 
+    #Make it persistant, update: sudo nano /etc/hostname
+    if [[ $HOSTNAME == pats-proto* ]];
+    then
+      echo "Hostname changed detected, good."
+    else
+      echo "Change hostname before running this script!"
+      exit 1
+    fi
 fi
+
+pushd ~/dependencies
 
 # Install packages
 [ -f dependencies-packages.done ] || {
@@ -86,23 +99,24 @@ fi
 	touch opencv-3.4.2.done
 }
 
+if [[ $1 -eq 1 ]] ; then
+    [ -f git.done ] || {
+        # Configure git
+        git config --global push.default simple
+        git config --global user.email "${HOSTNAME}@pats.com"
+        git config --global user.name $HOSTNAME
 
-[ -f git.done ] || {
-	# Configure git
-	git config --global push.default simple
-	git config --global user.email "${HOSTNAME}@pats.com"
-	git config --global user.name $HOSTNAME
-
-	#add deploy key
-	tar -xf pats_ssh_files.tar.xz
-	mkdir -p ~/.ssh
-	mv authorized_keys ~/.ssh/
-	mv id_rsa ~/.ssh/
-	chmod 600 ~/.ssh/id_rsa
-	mv id_rsa.pub ~/.ssh/
-	mv known_hosts ~/.ssh/
-	ssh-add ~/.ssh/id_rsa
-}
+        #add deploy key
+        tar -xf pats_ssh_files.tar.xz
+        mkdir -p ~/.ssh
+        mv authorized_keys ~/.ssh/
+        mv id_rsa ~/.ssh/
+        chmod 600 ~/.ssh/id_rsa
+        mv id_rsa.pub ~/.ssh/
+        mv known_hosts ~/.ssh/
+        ssh-add ~/.ssh/id_rsa
+    }
+fi
 
 # Install the Pats code
 [ -f pats_code.done ] || {
@@ -126,51 +140,57 @@ fi
 # Add to groups
 sudo usermod -a -G dialout $USER
 
-# Create nice symlinks
-[ -f symlinks.done ] || {
-	[ -f ~/.screenrc ] && {
-		cp -n ~/.screenrc{,.bak}
-		rm ~/.screenrc
-	}
-	ln -s ~/code/pats/config/.screenrc ~/
-	
-	[ -f ~/.bashrc ] && {
-		cp -n ~/.bashrc{,.bak}
-		rm ~/.bashrc
-	}
-	ln -s ~/code/pats/config/.bashrc ~/
+if [[ $1 -eq 1 ]] ; then
+    # Create nice symlinks
+    [ -f symlinks.done ] || {
+        [ -f ~/.screenrc ] && {
+            cp -n ~/.screenrc{,.bak}
+            rm ~/.screenrc
+        }
+        ln -s ~/code/pats/config/.screenrc ~/
+        
+        [ -f ~/.bashrc ] && {
+            cp -n ~/.bashrc{,.bak}
+            rm ~/.bashrc
+        }
+        ln -s ~/code/pats/config/.bashrc ~/
 
-	[ -f /etc/ssh/sshd_config ] && {
-		sudo cp  -n /etc/ssh/sshd_config{,.bak}
-		sudo rm /etc/ssh/sshd_config
-	}
-	sudo ln -s ~/code/pats/config/sshd_config /etc/ssh/
+        [ -f /etc/ssh/sshd_config ] && {
+            sudo cp  -n /etc/ssh/sshd_config{,.bak}
+            sudo rm /etc/ssh/sshd_config
+        }
+        sudo ln -s ~/code/pats/config/sshd_config /etc/ssh/
 
-	[ -f /etc/rc.local ] && {
-		sudo cp  -n /etc/rc.local{,.bak}
-		sudo rm /etc/rc.local
-	}
-	
-	rm ~/.ssh/config -f
-	ln -s ~/code/pats/config/sshconfig ~/.ssh/config
-	
-	sudo ln -s ~/code/pats/config/rc.local /etc/rc.local
+        [ -f /etc/rc.local ] && {
+            sudo cp  -n /etc/rc.local{,.bak}
+            sudo rm /etc/rc.local
+        }
+        
+        rm ~/.ssh/config -f
+        ln -s ~/code/pats/config/sshconfig ~/.ssh/config
+        
+        sudo ln -s ~/code/pats/config/rc.local /etc/rc.local
 
-	sudo ln -s ~/code/pats/config/45-pats_mm.rules /lib/udev/rules.d/
-	
-	echo "alias df='df -h -x squashfs -x tmpfs -x devtmpfs'" >> ~/.bash_aliases
+        
+        
+        echo "alias df='df -h -x squashfs -x tmpfs -x devtmpfs'" >> ~/.bash_aliases
+        
+        sudo systemctl restart ssh.service
+        
+        touch symlinks.done
+    }
+fi
 
-	touch symlinks.done
+#install driver for multi module:
+[ -f mm_install.done ] || {
+    sudo ln -s ~/code/pats/config/45-pats_mm.rules /lib/udev/rules.d/
+    sudo udevadm control --reload-rules && udevadm trigger
+    touch mm_install.done
 }
-
 popd
 
 sudo apt-get autoremove -y
 sudo apt-get clean -y
-
-#ssh-copy-id mavlab-gpu
-sudo systemctl restart ssh.service
-sudo udevadm control --reload-rules && udevadm trigger
 
 set +x
 echo "***********************************************************"
